requirements.txt
tensorflow
tensorflow-hub
faiss-cpu
scikit-image
streamlit
streamlit_cropper
matplotlib
tqdm
opencv-python
gdown